Feat: follower 컴포넌트에서 Axios API 호출 코드 분리
